backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic>=0.40.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
